update tests, separate fixtures
